Update README.md
Electronic Communication has become indispensable in this day and age thus the demand for
highly reliable mobile broadband services and cost- effective communication technology is
palpable. 4G LTE and its ability to provide fast, multimedia mobile data services materializes
the requisites of the hour. In this scenario, there exists a need for an effective simulator that
can mimic the end user environment and simulate the network to be deployed in the real
world. LTE Network simulation involves three stages which are Signal Generation, Resource
Grid Visualization and Characteristic Analysis. The project focuses on second stage of an
LTE network simulator, all the simulations and coding are done using JavaScript in Eclipse
IDE. The proposed resource grid displays the influence of various parameters on it and
provide information about the available network resources dynamically. Resource grid
Visualization in combination with Signal generation aids in analysis of various characteristics
of LTE network. This helps the service provider save revenue and ensure reliable
communications to user